undo changes in reader.rs
