--- src/config/reader/mod.rs
@@ -1,7 +1,0 @@
-pub mod reader;
-#[cfg(feature = "default")]
-mod reader_cli;
-#[cfg(not(feature = "default"))]
-mod reader_wasm;
-
-pub use reader::*;
